+++ src/struct_writer/default_template_rust.py
@@ -1,0 +1,9 @@
+import tomllib
+from pathlib import Path
+
+
+def default_template():
+    template = Path(__file__).parent / "../../examples/template_rust.toml"
+    template = template.absolute()
+    template = tomllib.loads(template.read_text("utf-8"))
+    return template